+++ nox/src/node.rs
@@ -159,6 +159,8 @@ impl<RT: AquaRuntime> Node<RT> {
             builtins_peer_id,
             config.node_config.default_service_memory_limit,
             config.node_config.allowed_effectors.clone(),
+            config.node_config.dev_mode_config.binaries.clone(),
+            config.node_config.dev_mode_config.enable,
         )
         .expect("create services config");
 
@@ -716,6 +718,10 @@ mod tests {
             to_peer_id(&config.root_key_pair.clone().into()),
             config.dir_config.air_interpreter_path.clone(),
             None,
+            None,
+            None,
+            None,
+            false,
         );
         let data_store_config = DataStoreConfig::new(config.dir_config.avm_base_dir.clone());